moved residuals to new module
